[MIG] stock_lot_scrap: Migration to 12.0
