+++ stock_lot_scrap/__manifest__.py
@@ -1,12 +1,13 @@
 # Copyright 2016 Carlos Dauden <carlos.dauden@tecnativa.com>
 # Copyright 2016 Pedro M. Baeza <pedro.baeza@tecnativa.com>
 # Copyright 2017 David Vidal <david.vidal@tecnativa.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
 {
     'name': 'Scrap Production Lot',
     'summary': 'This module adds a button in Production Lot/Serial Number '
                'view form to Scrap all products contained.',
-    'version': '11.0.1.0.0',
+    'version': '12.0.1.0.0',
     'category': 'Stock',
     'license': 'AGPL-3',
     'author': 'Tecnativa,'
