+++ stock_lot_scrap/i18n/fr.po
@@ -20,7 +20,7 @@ msgstr ""
 "Plural-Forms: nplurals=2; plural=(n > 1);\n"
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:18
+#: code:addons/stock_lot_scrap/models/stock_scrap.py:16
 #, python-format
 msgid "Lot was scrapped by <b>%s</b>."
 msgstr ""
@@ -31,27 +31,23 @@ msgid "Lot/Serial"
 msgstr "Lot/n° de série"
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:50
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:39
+#: model:ir.model,name:stock_lot_scrap.model_stock_scrap
 #, python-format
 msgid "Scrap"
 msgstr "Rebut"
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:76
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:65
 #, python-format
 msgid "This lot doesn't contain any quant in internal location."
 msgstr ""
 "Ce lot ne contient aucune quantité correspondant à l'emplacement interne."
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:48
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:37
 #, python-format
 msgid "This will scrap the whole lot. Are you sure you want to continue?"
 msgstr ""
 "Cela va mettre au rebut tout le lot. Etes - vous certains de vouloir "
 "continuer?"
-
-#. module: stock_lot_scrap
-#: model:ir.model,name:stock_lot_scrap.model_stock_scrap
-msgid "stock.scrap"
-msgstr ""
