+++ stock_lot_scrap/i18n/pt_BR.po
@@ -1,50 +1,49 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# * stock_lot_scrap
+#	* stock_lot_scrap
 #
-# Translators:
-# OCA Transbot <transbot@odoo-community.org>, 2017
-# enjolras <yo@miguelrevilla.com>, 2017
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 10.0\n"
+"Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2018-02-28 03:51+0000\n"
-"PO-Revision-Date: 2018-02-28 03:51+0000\n"
-"Last-Translator: enjolras <yo@miguelrevilla.com>, 2017\n"
-"Language-Team: Spanish (https://www.transifex.com/oca/teams/23907/es/)\n"
-"Language: es\n"
+"PO-Revision-Date: 2020-06-23 16:19+0000\n"
+"Last-Translator: Fernando Colus <fcolus1@gmail.com>\n"
+"Language-Team: none\n"
+"Language: pt_BR\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+"Plural-Forms: nplurals=2; plural=n > 1;\n"
+"X-Generator: Weblate 3.10\n"
 
 #. module: stock_lot_scrap
 #: code:addons/stock_lot_scrap/models/stock_scrap.py:16
 #, python-format
 msgid "Lot was scrapped by <b>%s</b>."
-msgstr ""
+msgstr "Lote foi marcado como sucata por <b>%s</b>."
 
 #. module: stock_lot_scrap
 #: model:ir.model,name:stock_lot_scrap.model_stock_production_lot
 msgid "Lot/Serial"
-msgstr "Lote/Serie"
+msgstr "Lote/Serial"
 
 #. module: stock_lot_scrap
 #: code:addons/stock_lot_scrap/models/stock_production_lot.py:39
 #: model:ir.model,name:stock_lot_scrap.model_stock_scrap
 #, python-format
 msgid "Scrap"
-msgstr "Desechar"
+msgstr "Sucata"
 
 #. module: stock_lot_scrap
 #: code:addons/stock_lot_scrap/models/stock_production_lot.py:65
 #, python-format
 msgid "This lot doesn't contain any quant in internal location."
-msgstr "Este lote no contiene ninguna cantidad en ubicación interna."
+msgstr "Este lote não contém qualquer quantidade em uma localização interna."
 
 #. module: stock_lot_scrap
 #: code:addons/stock_lot_scrap/models/stock_production_lot.py:37
 #, python-format
 msgid "This will scrap the whole lot. Are you sure you want to continue?"
-msgstr "Se desechará todo el lote. ¿Desea continuar?"
+msgstr ""
+"Isso vai colocar todo o lote como sucata. Você tem certeza que quer "
+"continuar?"
