+++ stock_lot_scrap/i18n/stock_lot_scrap.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 11.0\n"
+"Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
@@ -14,7 +14,7 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:18
+#: code:addons/stock_lot_scrap/models/stock_scrap.py:16
 #, python-format
 msgid "Lot was scrapped by <b>%s</b>."
 msgstr ""
@@ -25,25 +25,21 @@ msgid "Lot/Serial"
 msgstr ""
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:50
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:39
+#: model:ir.model,name:stock_lot_scrap.model_stock_scrap
 #, python-format
 msgid "Scrap"
 msgstr ""
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:76
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:65
 #, python-format
 msgid "This lot doesn't contain any quant in internal location."
 msgstr ""
 
 #. module: stock_lot_scrap
-#: code:addons/stock_lot_scrap/models/stock_scrap_lot.py:48
+#: code:addons/stock_lot_scrap/models/stock_production_lot.py:37
 #, python-format
 msgid "This will scrap the whole lot. Are you sure you want to continue?"
-msgstr ""
-
-#. module: stock_lot_scrap
-#: model:ir.model,name:stock_lot_scrap.model_stock_scrap
-msgid "stock.scrap"
 msgstr ""
