+++ stock_lot_scrap/models/__init__.py
@@ -1,1 +1,2 @@
-from . import stock_scrap_lot
+from . import stock_production_lot
+from . import stock_scrap
